Fix inter-packet delay bug
Enumeration test is working again, at least on the random Bluetooth dongle I tried.

The issue before was that we only added a single idle cycle after EOP, assuming
that there would be plenty of additional delay introduced by code overhead.
This is true (and necessary for performance) in rx, but in tx it violates spec
and usually overwhelms the device. With this patch, we correctly space out the
multiple output packets, while still receiving data from the device correctly.